requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
boto3==1.21.11
botocore==1.24.11
celery==5.2.6
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-seed==0.2.2
django-storages==1.12.3
django-timezone-field==4.2.1
djangorestframework==3.13.1
Faker==13.3.0
idna==3.3
install==1.3.4
jmespath==0.10.0
kombu==5.2.3
numexpr==2.8.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
passlib==1.7.4
prompt-toolkit==3.0.22
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.7
pytz==2021.3
redis==4.1.4
requests==2.27.1
s3transfer==0.5.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
supervisor==4.2.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.63.0
twilio==7.10.0
urllib3==1.26.7
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
XlsxWriter==3.0.3
